refactor(frontend): split playtest image prop runtime
